+++ src/visible_list.h
@@ -11,13 +11,13 @@ public:
 	visible_list() { }
 	visible_list(std::vector<visible*> objects) : objects(objects) { }
 
-	virtual bool intersects(const ray& r, float dist_min, float dist_max, collision& data) const;
+	virtual bool intersects(const ray& r, double dist_min, double dist_max, collision& data) const;
 };
 
-bool visible_list::intersects(const ray& r, float dist_min, float dist_max, collision& data) const {
+bool visible_list::intersects(const ray& r, double dist_min, double dist_max, collision& data) const {
 	collision temp_collision;
 	bool collision_occured = false;
-	float closest_dist = dist_max;
+	double closest_dist = dist_max;
 
 	for(auto obj : objects) {
 		if(obj->intersects(r, dist_min, closest_dist, temp_collision)) {
